Fixing build issues; fixing erroneous network error modal
