+++ src/Components/SettingsMenu.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { IonLabel, IonItem, IonListHeader, IonRadioGroup, IonList, IonRadio } from "@ionic/react";
 import { SegmentChangeEventDetail } from "@ionic/core";
 import { ServerEndpoint } from "../Utils/Types";
-import { config } from "../Utils/Constants";
+import { config } from "../Interfaces";
 
 require("dotenv");
