+++ src/Interfaces/Common.types.ts
@@ -1,0 +1,13 @@
+export enum SearchParams {
+  Domain = "domain",
+  NewName = "name",
+}
+
+export enum ServerAddress {
+  Localhost = "http://localhost:3002",
+  Production = "https://drinkentory-server.herokuapp.com",
+}
+
+export enum config {
+  SERVER_ENDPOINT_KEY = "serverEndpoint",
+}
