+++ src/Utils/Constants.ts
@@ -1,15 +1,1 @@
-export enum config {
-  SERVER_ENDPOINT_KEY = "serverEndpoint",
-}
-
 export const greyFont = "rgb(66, 66, 66)";
-
-export enum SearchParams {
-  Domain = "domain",
-  NewName = "name",
-}
-
-export enum ServerAddress {
-  Localhost = "http://localhost:3002",
-  Production = "https://drinkentory-server.herokuapp.com",
-}
